api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
apscheduler==3.10.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
croniter==1.4.1
pytz==2023.3
